refactor: moved the particular groups of functionality into their own files
