+++ Python/HAT/graph.py
@@ -1,6 +1,6 @@
 import numpy as np
 import scipy as sp
-import hypergraph
+from HAT import hypergraph
 
 
 """
