+++ Python/HAT/hypergraph.py
@@ -180,213 +180,6 @@ class Hypergraph:
         return Hypergraph(IM)
 
     
-    def laplacianMatrix(self, type='Bolla'):
-        """This function returns a version of the higher order Laplacian matrix of the hypergraph.
-
-        :param type: Indicates which version of the Laplacin matrix to return. It defaults to ``Bolla`` [1], but ``Rodriguez`` [2,3] and ``Zhou`` [4] are valid arguments as well.
-        :type type: str, optional
-        :return: Laplacian matrix
-        :rtype: *ndarray*
-
-        Several version of the hypergraph Laplacian are defined in [1-4]. These aim to capture
-        the higher order structure as a matrix. This function serves as a wrapper to call functions
-        that generate different specific Laplacians (See ``bollaLaplacian()``, ``rodriguezLaplacian()``,
-        and ``zhouLaplacian()``).
-
-        References
-        ----------
-        .. [1] Bolla, M. (1993). Spectra, euclidean representations and clusterings of hypergraphs. Discrete Mathematics, 117. 
-            https://www.sciencedirect.com/science/article/pii/0012365X9390322K
-        .. [2] Rodriguez, J. A. (2002). On the Laplacian eigenvalues and metric parameters of hypergraphs. Linear and Multilinear Algebra, 50(1), 1-14.
-            https://www.tandfonline.com/doi/abs/10.1080/03081080290011692
-        .. [3] Rodriguez, J. A. (2003). On the Laplacian spectrum and walk-regular hypergraphs. Linear and Multilinear Algebra, 51, 285–297.
-            https://www.tandfonline.com/doi/abs/10.1080/0308108031000084374
-        .. [4] Zhou, D., Huang, J., & Schölkopf, B. (2005). Beyond pairwise classification and clustering using hypergraphs. (Equation 3.3)
-            https://dennyzhou.github.io/papers/hyper_tech.pdf
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        if type == 'Bolla':
-            return self.bollaLaplacian()
-        elif type == 'Rodriguez':
-            return self.rodriguezLaplacian()
-        elif type == 'Zhou':
-            return self.zhouLaplacian()
-        
-    def bollaLaplacian(self):
-        """This function constructs the hypergraph Laplacian according to [1].
-
-        :return: Bolla Laplacian matrix
-        :rtype: *ndarray*
-
-        References
-        ----------
-        .. [1] Bolla, M. (1993). Spectra, euclidean representations and clusterings of hypergraphs. Discrete Mathematics, 117.
-               https://www.sciencedirect.com/science/article/pii/0012365X9390322K
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        dv = np.sum(self.IM, axis=1)
-        Dv = np.zeros((len(dv), len(dv)))
-        np.fill_diagonal(Dv, dv)
-        de = np.sum(self.IM, axis=0)
-        De = np.zeros((len(de), len(de)))
-        np.fill_diagonal(De, de)
-        DeInv = sp.linalg.inv(De, overwrite_a=True)
-        L = Dv - (self.IM @ DeInv @ self.IM.T)
-        return L
-        
-    def rodriguezLaplacian(self):
-        """This function constructs the hypergraph Laplacian according to [1, 2].
-
-        :return: Rodriguez Laplacian matrix
-        :rtype: *ndarray*
-
-        References
-        ----------
-        .. [1] Rodriguez, J. A. (2002). On the Laplacian eigenvalues and metric parameters of hypergraphs. Linear and Multilinear Algebra, 50(1), 1-14.
-               https://www.tandfonline.com/doi/abs/10.1080/03081080290011692
-        .. [2] Rodriguez, J. A. (2003). On the Laplacian spectrum and walk-regular hypergraphs. Linear and Multilinear Algebra, 51, 285–297.
-               https://www.tandfonline.com/doi/abs/10.1080/0308108031000084374
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        A = self.IM @ self.IM.T
-        np.fill_diagonal(A, 0)
-        L = np.diag(sum(A)) - A
-        return L
-    
-    def zhouLaplacian(self):
-        """This function constructs the hypergraph Laplacian according to [1].
-
-        :return: Zhou Laplacian matrix
-        :rtype: *ndarray*
-
-        References
-        ----------
-        .. [1] Zhou, D., Huang, J., & Schölkopf, B. (2005). Beyond pairwise classification and clustering using hypergraphs. (Equation 3.3)
-               https://dennyzhou.github.io/papers/hyper_tech.pdf
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        Dvinv = np.diag(1/np.sqrt(self.IM @ self.edgeWeights))
-        Deinv = np.diag(1/np.sum(self.IM, axis=0))
-        W = np.diag(self.edgeWeights)
-        L = np.eye(len(self.IM)) - Dvinv @ self.IM @ W @ Deinv @ self.IM.T @ Dvinv
-        return L        
-        
-    def adjTensor(self):
-        """This constructs the adjacency tensor for uniform hypergraphs.
-        
-        :return: Adjacency Tensor
-        :rtype: *ndarray*
-
-        The adjacency tensor :math:`A` of a :math:`k-`order hypergraph :math:`H` is the multi-way, hypergraph analog of the pairwise, graph
-        adjacency matrix. It is defined as a :math:`k-` mode tensor ( :math:`k-` dimensional matrix):
-
-        .. math::
-            A \in \mathbf{R}^{ \overbrace{n \\times \dots \\times n}^{k \\text{ times}}} \\text{ where }{A}_{j_1\dots j_k} = \\begin{cases} \\frac{1}{(k-1)!} & \\text{if }(j_1,\dots,j_k)\in {E}_h \\\\ 0 & \\text{otherwise} \end{cases},
-
-        as found in equation 8 of [1].
-
-        References
-        ==========
-        .. [1] C. Chen and I. Rajapakse, Tensor Entropy for Uniform Hypergraphs, IEEE TRANSACTIONS ON NETWORK SCIENCE AND ENGINEERING (2020)
-            (Equation 8) https://arxiv.org/pdf/1912.09624.pdf
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        order = int(sum(self.IM[:,0]))
-        denom = np.math.factorial(order - 1)
-        modes = len(self.IM) * np.ones(order)
-        A = np.zeros(modes.astype(int))
-        for e in range(len(self.IM[0])):
-            vxc = np.where(self.IM[:, e] == 1)[0]
-            vxc = list(permutations(vxc))
-            for p in vxc:
-                A[p] = 1 / denom
-        return A
-    
-    def degreeTensor(self):
-        """This constructs the degree tensor for uniform hypergraphs.
-        
-        :return: Degree Tensor
-        :rtype: *ndarray*
-
-        The degree tensor :math:`D` is the hypergraph analog of the degree matrix. For a :math:`k-` order hypergraph
-        :math:`H=(V,E)` the degree tensor :math:`D` is a diagonal supersymmetric tensor defined
-
-        .. math::
-            D \in \mathbf{R}^{ \overbrace{n \\times \dots \\times n}^{k \\text{ times}}} \\text{ where }{D}_{i\dots i} = degree(v_i) \\text{ for all } v_i\in V
-
-        References
-        ----------
-        .. [1] C. Chen and I. Rajapakse, Tensor Entropy for Uniform Hypergraphs, IEEE TRANSACTIONS ON NETWORK SCIENCE AND ENGINEERING (2020)
-            https://arxiv.org/pdf/1912.09624.pdf
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        order = int(sum(self.IM[:,0]))
-        modes = len(self.IM) * np.ones(order)
-        D = np.zeros(modes.astype(int))
-        for vx in range(len(self.IM)):
-            D[tuple((np.ones(order) * vx).astype(int))] = sum(self.IM[vx])
-        return D
-    
-    def laplacianTensor(self):
-        """This constructs the Laplacian tensor for uniform hypergraphs.
-        
-        :return: Laplcian Tensor
-        :rtype: *ndarray*
-
-        The Laplacian tensor is the tensor analog of the Laplacian matrix for graphs, and it is
-        defined equivalently. For a hypergraph :math:`H=(V,E)` with an adjacency tensor :math:`A`
-        and degree tensor :math:`D`, the Laplacian tensor is
-
-        .. math::
-            L = D - A
-
-        References
-        ==========        
-        .. [1] C. Chen and I. Rajapakse, Tensor Entropy for Uniform Hypergraphs, IEEE TRANSACTIONS ON NETWORK SCIENCE AND ENGINEERING (2020)
-            (Equation 9) https://arxiv.org/pdf/1912.09624.pdf
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        D = self.degreeTensor()
-        A = self.adjTensor()
-        L = D - A
-        return L
-        
-    def tensorEntropy(self):
-        """Computes hypergraph entropy based on the singular values of the Laplacian tensor.
-        
-        :return: tensor entropy
-        :rtype: *float*
-
-        Uniform hypergraph entropy is defined as the entropy of the higher order singular
-        values of the Laplacian matrix [1].
-        
-        References
-        ----------
-        .. [1] C. Chen and I. Rajapakse, Tensor Entropy for Uniform Hypergraphs, IEEE TRANSACTIONS
-             ON NETWORK SCIENCE AND ENGINEERING (2020) (Definition 7, Algorithm 1) 
-             https://arxiv.org/pdf/1912.09624.pdf
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        L = self.laplacianTensor()
-        _, S, _ = mla.hosvd(L, M=False, uniform=True)
-        return sp.stats.entropy(S)
-    
     def matrixEntropy(self, type='Rodriguez'):
         """Computes hypergraph entropy based on the eigenvalues values of the Laplacian matrix.
 
@@ -426,63 +219,6 @@ class Hypergraph:
         G = self.cliqueGraph()
         avgDist = nx.average_shortest_path_length(G)
         return avgDist
-    
-    def ctrbk(self, inputVxc):
-        """Compute the reduced controllability matrix for :math:`k-` uniform hypergraphs.
-
-        :param inputVxc: List of vertices that may be controlled
-        :type inputVxc: *ndarray*
-        :return: Controllability matrix
-        :rtype: *ndarray*
-
-        References
-        ==========
-        .. [1] Chen C, Surana A, Bloch A, Rajapakse I. "Controllability of Hypergraphs."
-            IEEE Transactions on Network Science and Engineering, 2021. https://drive.google.com/file/d/12aReE7mE4MVbycZUxUYdtICgrAYlzg8o/view
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        A = self.adjTensor()
-        modes = A.shape
-        n = modes[0]
-        order = len(modes)
-        Aflat = np.reshape(A, (n, n**(order-1)))
-        ctrbMatrix = self.bMatrix(inputVxc)
-        j = 0
-        while j < n and np.linalg.matrix_rank(ctrbMatrix) < n:
-            kprod = mla.kronExponentiation(ctrbMatrix, len(modes)-1)
-            nextCtrbMatrix = Aflat @ kprod;
-            ctrbMatrix = np.concatenate((ctrbMatrix, nextCtrbMatrix), axis=1)
-            r = np.linalg.matrix_rank(ctrbMatrix)
-            U, _, _ = sp.linalg.svd(ctrbMatrix)
-            ctrbMatrix = U[:,0:r]
-            j += 1
-        return ctrbMatrix
-        
-    def bMatrix(self, inputVxc):
-        """Constructs controllability :math:`B` matrix commonly used in the linear control system
-        
-        .. math::
-            \\frac{dx}{dt} = Ax+Bu
-
-        :param inputVxc: a list of input control nodes
-        :type inputVxc: *ndarray*
-        :return: control matrix
-        :rtype: *ndarray*
-
-        References
-        ==========
-        .. [1] Can Chen, Amit Surana, Anthony M Bloch, and Indika Rajapakse. Controllability of hypergraphs. IEEE Transactions
-            on Network Science and Engineering, 8(2):1646–1657, 2021. https://drive.google.com/file/d/12aReE7mE4MVbycZUxUYdtICgrAYlzg8o/view
-        """
-        # Auth: Joshua Pickard
-        #       jpic@umich.edu
-        # Date: Nov 30, 2022
-        B = np.zeros((len(self.nodeWeights), len(inputVxc)))
-        for i in range(len(inputVxc)):
-            B[inputVxc[i], i] = 1
-        return B
     
     def clusteringCoef(self):
         """Computes clustering average clustering coefficient of the hypergraph.
